+++ WebThingAdapter.h
@@ -10,4 +10,4 @@
 
 #pragma once
 
-#include "ESPWebThingAdapter.h"
+#include "adapter/ESPWebThingAdapter.h"
