+++ adapter/EthernetWebThingAdapter.h
